Add files via upload
The Realtime Election Voting System is a scalable and interactive application designed to simulate a live election voting process. This system processes, analyzes, and visualizes votes in real time using a robust architecture built with:

Python: Core logic and data processing.
Apache Kafka: For handling real-time streaming of voting data.
Apache Spark Streaming: For enriching, aggregating, and processing votes on the fly.
PostgreSQL: For storing and managing structured data such as candidates, voters, and vote logs.
Streamlit: To create a user-friendly dashboard for monitoring the election results in real time.
Docker Compose: For seamless deployment and containerization of all services.
With this system, users can observe a dynamic voting process, showcasing data engineering and real-time analytics in action. It is an excellent demonstration of modern data pipelines and distributed systems.